+++ src/main/java/spacewars/view/menus/how_to_play/ArrowsViewer.java
@@ -13,7 +13,7 @@ public class ArrowsViewer {
         this.image = imageLoader.get("images/arrows_how_to_play.png");
     }
 
-    public void draw(GUI gui, int x, int y) throws IOException {
+    public void draw(GUI gui, double x, double y) throws IOException {
         image.draw(gui, x, y);
     }
 }
